+++ saleor/core/utils/translations.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, Tuple, Union
 
+from django.conf import settings
 from django.db import models
-from django.utils.translation import get_language
 
 
 class TranslationWrapper:
@@ -27,12 +27,6 @@ class TranslationWrapper:
         return str(instance)
 
 
-class TranslationProxy:
-    def __get__(self, instance, owner):
-        locale = get_language()
-        return TranslationWrapper(instance, locale)
-
-
 class Translation(models.Model):
     language_code = models.CharField(max_length=35)
 
@@ -51,3 +45,9 @@ class Translation(models.Model):
 
     def get_translation_context(self) -> Dict[str, Any]:
         return {}
+
+
+def get_translation(instance, language_code=None) -> TranslationWrapper:
+    if not language_code:
+        language_code = settings.LANGUAGE_CODE
+    return TranslationWrapper(instance, language_code)