+++ saleor/graphql/product/types/products.py
@@ -840,7 +840,10 @@ class Product(ChannelContextTypeWithMetadata[models.Product]):
     )
     is_available = graphene.Boolean(
         address=destination_address_argument,
-        description="Whether the product is in stock and visible or not.",
+        description=(
+            "Whether the product is in stock, set as available for purchase in the "
+            "given channel, and published."
+        ),
     )
     tax_type = graphene.Field(
         TaxType,
@@ -934,7 +937,12 @@ class Product(ChannelContextTypeWithMetadata[models.Product]):
         description="Date when product is available for purchase."
     )
     is_available_for_purchase = graphene.Boolean(
-        description="Whether the product is available for purchase."
+        description=(
+            "Refers to a state that can be set by admins to control whether a product "
+            "is available for purchase in storefronts. This does not guarantee the "
+            "availability of stock. When set to `False`, this product is still visible "
+            "to customers, but it cannot be purchased."
+        )
     )
     tax_class = PermissionsField(
         TaxClass,